requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
audioread==3.0.1
blinker==1.8.2
bottle==0.13.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
docx==0.2.4
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.9.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
nibabel==5.3.1
nipype==1.8.6
noisereduce==3.0.2
numba==0.60.0
numpy==2.0.2
openai==1.52.0
openai-whisper==20240930
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.2
pooch==1.8.2
prov==2.0.1
proxy_tools==0.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pydub==0.25.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMuPDF==1.24.11
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pythonnet==3.0.4
pytz==2024.2
pywebview==5.2
pywin32-ctypes==0.2.3
pyxnat==1.6.2
rdflib==6.3.2
regex==2024.7.24
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
semantic-version==2.10.0
setuptools-rust==1.10.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.0.4
